requirements.txt
azure.cognitiveservices.speech
chromadb
grpcio
keyboard
openai
numpy
protobuf==3.20.1
python-dotenv
websocket-client
websockets
tiktoken
SpeechRecognition
tokenizers
elevenlabs
flask
Flask-MySQLdb
langchain_openai
langchain
langchain_community
pydub
flask-socketio
flask_sqlalchemy
config
flask_migrate
beautifulsoup4
lxml
lxml_html_clean
readability
readability-lxml
